requirements.txt
amqp==2.6.0
asgiref==3.2.10
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.1.1
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-rest-auth==1.1.0
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-firebase-authentication==1.0.9
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-phone-verify==2.0.0
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-rest-firebase-auth==0.2.0
django-wkhtmltopdf==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-util==0.0.27
drf-yasg==1.17.1
elasticsearch==7.8.0
firebase-admin==4.3.0
flower==0.9.5
future==0.18.2
gcloud==0.17.0
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.8.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
gunicorn==20.0.4
httplib2==0.18.1
humanize==2.5.0
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
joblib==0.16.0
jsonfield==3.1.0
jsonschema==3.2.0
jws==0.1.3
kiwisolver==1.2.0
kombu==4.6.11
MarkupSafe==1.1.1
matplotlib==3.3.0
msgpack==1.0.0
nexmo==2.4.0
numpy==1.19.1
oauth2client==3.0.0
oauthlib==3.1.0
packaging==20.4
pandas==1.1.0
pdfkit==0.6.1
phonenumbers==8.12.7
Pillow==7.2.0
prometheus-client==0.8.0
protobuf==3.12.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
PyJWT==1.7.1
pyngrok==4.1.6
pyparsing==2.4.7
Pyrebase==3.0.27
pyrsistent==0.16.0
pyTelegramBotAPI==3.7.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-jwt==2.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.3.1
swagger-spec-validator==2.7.3
swapper==1.1.2.post1
threadpoolctl==2.1.0
tornado==6.0.4
twilio==6.44.2
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
wget==3.2
